requirements.txt
Django==1.3
django-rosetta
django-modeltranslation
